refactor: renamed label file
